requirements.txt
Flask==0.10.1
Flask-Migrate==1.6.0
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
pymysql==0.7.11